refactor: delete task and move structures to models
The task will not be necessary after removing the `mech-client`.
Moreover, the structures that were moved in the models have been slightly modified so that they can better represent the mech's response.
